pyproject.toml
[project]
name = "nonebot-plugin-anime-notification"
version = "0.1.0"
description = "✨ 番剧更新推送 & 番剧信息获取 NoneBot 插件 ✨"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-plugin-alconna>=0.38.2",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-localstore>=0.6.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.28",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
